+++ src/main/java/com/ua/teamconnect/tracker/model/entity/projection/Anniversary.java
@@ -3,7 +3,7 @@ package com.ua.teamconnect.tracker.model.entity.projection;
 import java.time.LocalDate;
 
 public interface Anniversary {
-    Long getUserId();
+    Integer getUserId();
     String getFirstName();
     String getLastName();
     String getAvatarUrl();
